Added support for halft-stars as a minimum value.

+++ HCSStarRatingView/HCSStarRatingView.h
@@ -25,7 +25,7 @@
 IB_DESIGNABLE
 @interface HCSStarRatingView : UIControl
 @property (nonatomic) IBInspectable NSUInteger maximumValue;
-@property (nonatomic) IBInspectable NSUInteger minimumValue;
+@property (nonatomic) IBInspectable CGFloat minimumValue;
 @property (nonatomic) IBInspectable CGFloat value;
 @property (nonatomic) IBInspectable CGFloat spacing;
 @property (nonatomic) IBInspectable BOOL allowsHalfStars;
